# api.yaml
openapi: 3.0.0
info:
  title: Placement Portal Application API
  description: REST API for the Placement Portal Application (PPA V2) - App Dev II Jan 2026
  version: 2.0.0

servers:
  - url: http://127.0.0.1:5000
    description: Local development server

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

security:
  - BearerAuth: []

paths:

  # ── AUTH ──────────────────────────────────

  /api/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                  example: admin
                password:
                  type: string
                  example: admin123
      responses:
        '200':
          description: Login successful — returns JWT token + user info
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  role:
                    type: string
                    enum: [Admin, Company, Student]
                  user_id:
                    type: integer
                  username:
                    type: string
        '401':
          description: Invalid credentials
        '403':
          description: Account deactivated or pending admin approval

  /api/register:
    post:
      tags: [Auth]
      summary: Register a new Student or Company account
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password, role]
              properties:
                username:
                  type: string
                email:
                  type: string
                password:
                  type: string
                role:
                  type: string
                  enum: [Student, Company]
                cgpa:
                  type: number
                  description: Required for Student
                  example: 8.5
                company_name:
                  type: string
                  description: Required for Company
                description:
                  type: string
                hr_contact:
                  type: string
                website:
                  type: string
      responses:
        '201':
          description: Registered successfully
        '400':
          description: Invalid role

  # ── ADMIN ─────────────────────────────────

  /api/admin/dashboard_data:
    get:
      tags: [Admin]
      summary: Get dashboard stats, all students, companies and drives
      responses:
        '200':
          description: Dashboard data
          content:
            application/json:
              schema:
                type: object
                properties:
                  stats:
                    type: object
                    properties:
                      total_students:
                        type: integer
                      total_companies:
                        type: integer
                      total_drives:
                        type: integer
                  students:
                    type: array
                    items:
                      type: object
                  companies:
                    type: array
                    items:
                      type: object
                  drives:
                    type: array
                    items:
                      type: object

  /api/admin/search:
    get:
      tags: [Admin]
      summary: Search students, companies, or drives
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: role
          in: query
          required: true
          schema:
            type: string
            enum: [Student, Company, Drive]
      responses:
        '200':
          description: Matching results

  /api/admin/all_applications:
    get:
      tags: [Admin]
      summary: View all student applications across all drives
      responses:
        '200':
          description: Global applications list

  /api/admin/toggle_user_status/{user_id}:
    post:
      tags: [Admin]
      summary: Blacklist or re-activate a student or company user
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Status toggled
        '400':
          description: Cannot toggle Admin account

  /api/admin/approve_company/{comp_id}:
    post:
      tags: [Admin]
      summary: Approve a company registration
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Company approved
        '404':
          description: Company not found

  /api/admin/reject_company/{comp_id}:
    post:
      tags: [Admin]
      summary: Reject and permanently delete a company registration
      parameters:
        - name: comp_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Company rejected and removed
        '404':
          description: Company not found

  /api/admin/approve_drive/{drive_id}:
    post:
      tags: [Admin]
      summary: Approve a placement drive — makes it visible to students
      parameters:
        - name: drive_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Drive approved
        '404':
          description: Drive not found

  /api/admin/reject_drive/{drive_id}:
    post:
      tags: [Admin]
      summary: Reject a placement drive
      parameters:
        - name: drive_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Drive rejected
        '404':
          description: Drive not found

  # ── COMPANY ───────────────────────────────

  /api/company/dashboard/{user_id}:
    get:
      tags: [Company]
      summary: Get company details and all created drives with applicant counts
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Company dashboard data
        '404':
          description: Company not found

  /api/company/create_drive:
    post:
      tags: [Company]
      summary: Submit a new placement drive for admin approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, title, date, deadline]
              properties:
                user_id:
                  type: integer
                title:
                  type: string
                  example: Software Engineer
                description:
                  type: string
                eligibility:
                  type: number
                  description: Minimum CGPA required (null = open to all)
                  example: 7.5
                date:
                  type: string
                  format: date
                  example: "2026-04-15"
                deadline:
                  type: string
                  format: date
                  example: "2026-04-10"
      responses:
        '201':
          description: Drive created — pending admin approval
        '403':
          description: Company not yet approved by admin
        '404':
          description: Company profile not found

  /api/company/applications/{drive_id}:
    get:
      tags: [Company]
      summary: View all applicants for a specific drive
      parameters:
        - name: drive_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: List of applicants with CGPA, resume, status, interview info

  /api/company/update_status:
    post:
      tags: [Company]
      summary: Update application status for a student
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_id, status]
              properties:
                app_id:
                  type: integer
                status:
                  type: string
                  enum: [Shortlisted, Selected, Rejected]
      responses:
        '200':
          description: Status updated
        '404':
          description: Application not found

  /api/company/schedule_interview:
    post:
      tags: [Company]
      summary: Schedule an interview date and notes for an applicant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [app_id, interview_date]
              properties:
                app_id:
                  type: integer
                interview_date:
                  type: string
                  format: date
                  example: "2026-04-20"
                interview_notes:
                  type: string
                  example: "Google Meet link will be shared"
      responses:
        '200':
          description: Interview scheduled
        '404':
          description: Application not found

  # ── STUDENT ───────────────────────────────

  /api/student/dashboard/{user_id}:
    get:
      tags: [Student]
      summary: Get all approved drives with CGPA eligibility status
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Drive list with is_eligible and ineligibility_reason per drive

  /api/student/profile/{user_id}:
    get:
      tags: [Student]
      summary: Get student profile — email, CGPA, resume
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Student profile
        '404':
          description: User not found

  /api/student/history/{user_id}:
    get:
      tags: [Student]
      summary: Get complete placement application history
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Application history with interview details

  /api/student/apply:
    post:
      tags: [Student]
      summary: Apply for a placement drive (validates CGPA eligibility on backend)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [user_id, drive_id]
              properties:
                user_id:
                  type: integer
                drive_id:
                  type: integer
      responses:
        '200':
          description: Applied successfully
        '400':
          description: Already applied to this drive
        '403':
          description: Not eligible — CGPA below minimum requirement
        '404':
          description: Student or drive not found

  /api/student/update_profile:
    post:
      tags: [Student]
      summary: Update student email, CGPA and/or upload resume (multipart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                user_id:
                  type: integer
                email:
                  type: string
                cgpa:
                  type: number
                  example: 8.5
                resume:
                  type: string
                  format: binary
      responses:
        '200':
          description: Profile updated successfully
        '404':
          description: User not found

  /api/student/export-history:
    post:
      tags: [Student]
      summary: Trigger async Celery job to export application history as CSV via email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                  example: student@example.com
      responses:
        '202':
          description: Job queued — student receives email with CSV download link
        '400':
          description: Email is required
        '404':
          description: No user found for this email